add make-wrapper-procedure
A program can use `make-wrapper-procedure` to synthesize a function
that reports a given arity mask (without calling `compile`) or to
associate data with a procedure that can be extracted with
`wrapper-procedure-data`.

In addition, `set-wrapper-procedure!` suports modifying the
implementation of a synthesized procedure. Although similar
functionality could be achieved with `(lambda args (apply (unbox proc)
args))`, a wrapper procedure can dispatch to another procedure without
allocating a list for the arguments.

The interpreter uses an internal variant of procedure wrappers to
cooperate with `procedure-arity-mask`.