cptypes: add support for call-with-values, apply and similar functions
In some procedures, one of the arguments is a function that will surely be called
and the result is the result of the whole expression. These procedures need an
special version of define-specialize that gives more control.

Original commit: racket/ChezScheme@5224276a85e5e619a9c40454bc9621f4dad4784f