collect in main thread when active
For a collect rendezvous, call the collect-notify handler in
the main thread if it is active. A collect-notify handler can
then make sure the main thread is active and try again, if
that's useful to an application.

Original commit: racket/ChezScheme@fc0c54889b6d221bf8da8e4f0ffe9065769c778f