Fix makefile for MSYS2
When MSYS2 bash sees the command "cmd.exe /c make.bat" it will
silently convert the /c into c:\, as it translates unixy paths
to Windows paths. This can be disabled by setting the environment
variable MSYS2_ARG_CONV_EXCL to *. So do that for this particular
command. It will be harmless in other environments like Cygwin.