make `configure` work in a separate build directory
Adjust `configure` and makefiles so that `configure` can be run in any
directory, not just the source directory. Initial boot files are found
at either the current directory or the source directory.

Also, add support for `-B` and `--Boot` flags to accept a path that is
resolved relative to the current directory, instead of relative to the
installation.

Original commit: racket/ChezScheme@53b80fe7f42b287c687b7eb1f7cc12231f4bdf49