adjust GC to accommodate a 0-sized continuation stack segment
A 0-sized continuation stack segment can show up when a continuation
is created only because the current one does not record attachments;
that happens with a continuation is created to split the stack of a
large continuation. We would change things to initialize the
attachments field instead of creating an extra continuation, but this
patch takes the approach of allowing a zero-sized stack segment. The
GC needs to avoid allocating a 0-sized stack segment, because the
`find_room` macro expects a non-zero size.