fix hash tables for certain likely fixnum distibutions
A hash table only uses the lower few bits of a hash code, because it
masks the hash code by [one less than] the power-of-two size of the
bucket array. That truncation interacts badly with the hashing
function for fixnums, which is the identity function; if the the lower
several bits of the fixnum stay the same for many keys and the upper
bits change, then there are many hash collisions --- and that's a
relatively likely distribution.

Fix the hash-table implementation by mixing the hash code to let
higher bits influence the lower bits: xor the high half of the bits
with the lower half (which doesn't lose information, because xoring
again would recover the original number), then do the same for the
high one-fourth of bits in the low half, and then (on a 64-bit
platform) the high one-eighth of the low one-fourth of the bits.

Original commit: racket/ChezScheme@d5fce7cd76881e3022648804ad548c41d96dcc43