x86_64: fix corner case of foreign calls for signed 8-bit or 16-bit integers
For x86_64, it seems that integer values passed in registers should be
sign-extended to at least 32 bits. (Clang on macOS apparently only
takes advantage of the sign extension when compiling with
optimization.) In the case that a foreign-call argument is passed via
`(& integer-8)` or `(& integer-16)` --- which no one is likely to do
in practice --- the sign extension didn't happen as it should.