move ptr-tagging assumption out of rktboot
Add `record-ptr-offset` so that rktboot can use that instead of
assuming an offest of 1.

Also, introduce `type-untyped` and use it instead of `typemod` when
the intent is to leave an address unchanged by tagging. That makes the
implementation a little clearer, and it reduces the code that would
have to be changed to modify the tagging discipline (e.g., to tag by
adding instead of subtracting from an address).

Original commit: racket/ChezScheme@11ca54ed8ac05ad19fdc7b715705cb20de71722e