xterm-256color hack to improve portability of terminal on Linux
On libncurses6 systems, a statically linked libncurses5 is not happy
with an "xterm-color256" terminal, maybe because the host system's
terminfo database specifies things that libncurses5 does not support.
Avoid that problem by swapping "xterm" in place for that case.

There may be other terminal modes that could run into the same
problem, but the goal is just to support portable-as-possible Linux
executables that work in typical environments, and it's not a disaster
if terminal support fails. Building specifically on/for the host
system covers the rest.

Original commit: racket/ChezScheme@653ba761a514e9fe715d1efe05a9cad8995d7e52