Create compile-time values for base ftypes. (#569)
* Allow native-ftypes to work, even when a non-ftype is bound to the ftype name.

updated ftypes to allow native ftypes to be used even if another
non-ftype syntactic binding for the type exists.  for instance, if a
syntactic binding for integer-32 is introduced which is not bound to
an ftd, using integer-32 in the context of an ftype will still work.
this change also allows an ftype binding to be created for a native
type by using (define-ftype <native-type> <native-type>), allowing
users to create standard ftype bindings for the native-types if that
is preferred.
  s/ftype.ss

* Adding some tests for ftype changes.

Added a couple of tests for the ftype change to avoid having syntactic
bindings overlap the native ftype names without interfering with the
operation of ftype definitions.

This also includes a fix to the CSUG ftype section, thanks to Oscar
Waddell for the change.

* Small clarification on endianness in CSUG.

* Updated release notes to include note about ftype update.

* Added corrections suggested by @burgerrg

Original commit: 0747a5d5e4990d402f350291fa0f44270d0c4267