check if first part of seq is omitable in cptypes
If the first part of a seq has a potential side effect, cptypes may
prove that it's safe and omitable, and it will be removed in the second
pass of cp0.

Check it in cptypes too to remove it direcctly.

Original commit: racket/ChezScheme@382b71f41ac616f1a021b95e8b13e727912ab8f0