add libffi support for pb
By using libffi, we can fill in the one piece of Chez Scheme that was
previously unimplemented for pb (portable bytecode) mode: calling
foreign functions and converting a closure to a foreign function.

In principle, this change allows Racket CS to run on any platform
where Racket BC runs (in interpreted mode); the performance of
interpreted Racket code is about the same in those two. But since
Racket BC's compiler is in C and its expander can be compiled to C
(via cify), BC is still much faster to expand and compile programs. A
next step might be to set up a similar compiltion via C for the static
part of Chez Scheme and Racket.

Original commit: racket/ChezScheme@9351fae7e7ef09dababe55305032cbb0973b0dd6