Add arm64mac machine type and arm64 backend file
Add files for arm64mac.def and arm64.ss files in support of an arm64mac
machine-type for native Apple Silicon support. The actual machine-type name
may change.

This changeset includes complete draft implementations for sections 1 and 2
of the arm64 backend, i.e., define-registers and define-instructions. A
future changeset will include an implementation for the assembler.