fix exact/inexact comparisons where fixnum->flonum had loss of precision
makes <=, =, and >= transitive as required by R6RS by changing behavior
of mixed exact/inexact comparisons in the range where fixnums have
greater precision than flonums.  For example, the flonum representation
of 9007199254740992.0 and 9007199254740993.0 is identical,
but obviously the fixnums 9007199254740992 and 9007199254740993
are not.  In order for = and the inclusive inequalities to be transitive,
exact/inexact comparisons must behave as if the inexact number was replaced by
its exact representative.  A similar change was made for the strict
inequalities as well (even though they were already transitive).