add grapheme-cluster functions
The main function is `char-grapheme-step`, which supports streaming
characters through a counter. The `string-grapheme-span` function
builds on that to report the number of characters that continue a
grapheme cluster from a given starting point in a string, and
`string-grapheme-count` builds on that to count the number of grapheme
clusters in a (sub)string.

The `char-extended-pictographic?` and `char-grapheme-break-property`
functions expose extra character properties that
`char-grapheme-cluster-step` uses internally.

Original commit: racket/ChezScheme@1a9b5dd1fe8d2f8f75607eefa9a75f36f6ec2e12