fix Mac AArch64 alignment for on-stack FFI arguments
The step that looks at the next argument to determine padding should
look a the next argument that is on the stack. The immediately next
argument might be in a register.

Also, repair copying into the stack of records that are not a multiple
of the word size. The copy was reading the right bytes, but could end
up zeroing subsequent bytes that are already filled with an argument.

Closes racket/ChezScheme#50

Original commit: racket/ChezScheme@00c5789736914bdb34e5bfcfe95b9bce40853d76