Use Semantic Versioning for pre-release versions
Replace `scheme-build-number` (wrong terminology) with
`scheme-pre-release`. The result of `scheme-pre-release` is #f for a
release version, and it's a positive integer for a pre-release
version.

Since a #f pre-release is after a non-#f pre-release version with the
same major, minor, and patch numbers, we also bump the main version
here to 9.9.9 --- intended to stay clear of Chez Scheme releases in
the near term.

For consistency with Chez Scheme programs that have relied on the
correspondence between `scheme-version` and `scheme-version-number`,
this commit omits and pre-release number in the result of
`scheme-version` when called with 0 arguments. The pre-release
version, if any, is included in the result of `(scheme-version #t)`.

Original commit: racket/ChezScheme@b8be3f793c2d7d0951428e21b5e3cf44ae12e33d