fix size accounting when object in list twice
When an object is referenced twice in a list passed to
`compute-size-increments` or the 4th argument to `collect`,
the implementation implicitly assumed that all the objects
were distinct, but they might not be.

Also, fix the docs to properly account for the size of an emphemeron
pair in an example.

Thanks to @jtaylor-us for noticing the problems.

Original commit: racket/ChezScheme@93280691170babbf071b920b5407f9c39855f90d