add `fasl-read` check for external vector size
Also, refine interning and graph-structure references for integers.
The goal is to ensure that sharing of bignums and graph structure
that's explicit in a fasl file is consistent across 32-bit and 64-bit
platforms, including when cross-compiling from one to the other.

Original commit: racket/ChezScheme@08208878a2752d92d0a55d69505f437711e5b719