"mark-set" -> "marks"
Change procedure names that contain "mark-set", such as
`continuation-mark-set?`, to instead have just "marks", as in
`continuation-marks?`. This is less directly compatible with Racket,
but it avoids the term "set", which is misleading because it's not a
set data structure.

Original commit: racket/ChezScheme@37e48c1ad15f3e1245a2896a395fab455654fc33