find boot files relative to the executable by default (#755)
Building on recent support for "%x" everywhere, the change here makes
the default boot-file search path

 "%x:%x/../lib/csv%v/%m:%x/../../boot/%m"

which searches in the immediate directory, then in the place relative
to the executable's install location to reach installed boot files,
and then relative to the workarea executable to reach workarea boot
files. More precisely, the middle path in the default takes into
account the configured bin and lib target directories.

As a result of this change, it's much easier to run a just-built
executable before installing it, and installing to a non-system
directory produces something that you can move around in the
filesystem.

As a further step to make things work in the default way, `configure`
now supports `--prefix` as an alias for `--installprefix`.

The `--installabsolute` flag uses the old default search path, which
is absolute relative to the target installation directory.