improve some tests that have been unreliable (#767)
A `compute-size-increments` test has been a long-time problem. It
turns out that it was failing randomly due to a random list of symbols
stashed in a thread parameter by an earlier test; sometimes, one of
those symbols can reach a lot of data. The solution here is to
not count data reachable from thread parameters.

A `cost-center-time` test has been failing on very slow machines. As a
timing test, it's inherently unreliable, but turning up the limit
should make failures rare enough without making the test useless.