add `CFLAGS+=...` handling to `configure`
The `+=` argument form provides a way of extending the current or
default value of a compiler-configuration variable, instead of
replacing it. For example, `CFLAGS+=-std=gnu99` adds `-std=gnu99` to
the C compiler flags, instead of requiring `CFLAGS=-O2 -std=gnu99` to
keep the default and add a new flag.

In addition, changed `configure` to treat the existence of a `CFLAGS`
environment variable the same as if `CFLAGS=` is provided as an
argument.