Use GNU89 for kernel C
C99/GNU99 is obviously well established and a sensible choice of
standard for a C code base. Still, in some environments where it
remains useful to compile Chez Scheme, the default `gcc` mode is
GNU89.

Since C isn't the main development language for Chez Scheme, since
it's easy to stick to GNU89, and since that usefully reduces build
friction, this commit makes the few changes needs to get back to
GNU89. GitHub CI is configured to use `-std=gnu89` for two builds to
help maintain the choice.

GNU89 is a little more flexible than C89 (ANSI C), because it allows
declarations in the middle of a block, but it does not allow variables
defined in the parentheses after `for`. We also rely on GNU extensions
for inlining declarations.