adjust cp0 to combine `begin` rotation and variable dropping
A follow-up to c081296367, this commit adjusts the cp0 change to avoid
skipping the variable-dropping rewrite when the `begin` rotation
applies. This combination passes the new test, passes old tests with
small adjustments, and allows Racket to pass some tests that are
similar to "cp0.ms" tests.

Meanwhile, c081296367 should have noted the PR (#789) it squashes and
some author information that was lost in the squash:
Co-authored-by: R. Kent Dybvig <dyb@scheme.com>