Repairs and improvements for building with external dependencies (#807)
* configure: fix zlib and lz4 submodule checks

Conditionals to skip the submodule checks were using the wrong
variable names, so the checks were always skipped. The correct
behavior is to perform the check unless given `ZLIB=<zlib>` or
`LZ4=<lz4>`, as applicable.

* configure: support `ZUO=<zuo>`

Supplying `ZUO=<zuo>` skips the submodule check in `configure`
and configures the generated makefile not to build or remove Zuo.

* configure: support `STEXLIB=<stex>`

For compatibility with older scripts, when not explicitly configured,
continue to honor the `STEXLIB` environment variable at build time.