repair GC for reference bytevector whose length isn't a multiple of a word
A reference bytevector is intended to be used as a multiple of a word
size, but there's no explicit constraint, and so it should work with
any length. When the length is not a multiple of the word size, then
in a copying step, extra bytes at the end of the bytevector were
missed.