configure: Fix flags handling for pb little-endian targets
While upgrading the Alpine Linux ChezScheme package to 0.10.0, I
encountered a weird build failure on ppc64le (`--machine=tpb64l`)
were it fails to link against ncurses.

Further debugging revealed that this is due to the fact that the
`configure` script does not add `-lncurses` to the linker flags
if the machine type does not end in `*le`. Since 0.10.0 introduced
little-endian targets (such as `tpb64l`) which do not end in `*le`
I believe the `configure` script needs to be adjusted accordingly.

This patch fixes the described build failure on Alpine Linux.