configure: improve inference of compiler flags for pb builds
When building an pb variant like `tpb64l`, and even when building for
a platform without native-code support, the "configure" script may
still be able to infer suitable C compilation and linking flags for
the kernel. Add an `--os=...` flag that can be used to override that
inference in case it's not correct (as in the case of cross
compilation).