improve cross-compile instructions and makefile support
Add a `--cross` argument to `configure` to indicate that the generated
makefile's `make` should be like `make kernel`, and a pb-based
bootstrap should rerun `configure` to build for the host machine
instead of the specified target machine.

Also, add support for `DESTDIR` as a more conventional alternative to
the `--temproot` configure flag. While we're at it, recognize
`CFLAGS_FOR_BUILD` for building `bin/zuo`.

As part of the change to `configure`, having `CC_FOR_BUILD` set does
not by itself disable the use of pb for creating or updating boot
files. Depending on how boot files are obtained, `--force` now may be
needed when using `CC_FOR_BUILD` as part of a cross compile that
doesn't use `--cross`.